Improve installation instructions
Remove the reference to the kaitlinsalzke.com instructions for getting
started with Omni Automation because they contain a note saying that
they are out of date and refer the reader to a different document.

Update the instructions to match the current GitHub UI.

Reference the "Omni Automation: Installing Plug-Ins" document for how to
install the plug-in.